feat: la cantidad de conceptos admite decimales (horas) con coma
El input de cantidad pasa de type=number/parseInt a un campo de texto
que acepta coma o punto como separador decimal. Se normaliza a numero
para calculos y al guardar, y se muestra con coma en el form y el PDF.

+++ web/backend/src/database.js
@@ -27,7 +27,7 @@ db.exec(`
     id INTEGER PRIMARY KEY AUTOINCREMENT,
     factura_id TEXT,
     descripcion TEXT,
-    cantidad INTEGER,
+    cantidad REAL,
     precio_unitario REAL,
     FOREIGN KEY (factura_id) REFERENCES facturas(id) ON DELETE CASCADE
   )
